refactor: organize hookers into domain subpackages
