+++ app/src/main/java/io/github/twyora/douyinenhancer/hook/settings/SettingsHooker.kt
@@ -1,4 +1,4 @@
-package io.github.twyora.douyinenhancer.hook
+package io.github.twyora.douyinenhancer.hook.settings
 
 import android.app.Activity
 import android.content.Context
@@ -9,10 +9,12 @@ import com.highcapable.yukihookapi.hook.entity.YukiBaseHooker
 import com.highcapable.yukihookapi.hook.factory.injectModuleAppResources
 import com.highcapable.yukihookapi.hook.log.YLog
 import io.github.twyora.douyinenhancer.R
+import io.github.twyora.douyinenhancer.hook.DouyinPackage
+import io.github.twyora.douyinenhancer.hook.HookOnMainProcess
+import io.github.twyora.douyinenhancer.ui.SettingsDialog
 import io.github.twyora.douyinenhancer.utils.getField
 import io.github.twyora.douyinenhancer.utils.invokeMethod
 import io.github.twyora.douyinenhancer.utils.resolveMethod
-import io.github.twyora.douyinenhancer.ui.SettingsDialog
 
 @HookOnMainProcess
 object SettingsHooker : YukiBaseHooker() {
